+++ src/main/java/org/trash_hunter/trashes/Bottle.java
@@ -21,7 +21,7 @@ public class Bottle extends Trash {
         appearanceRangeYSup= 700-this.height;
     }
     public Bottle (){this(0,0);}
-    public Bottle (long id){
+    public Bottle (int id){
         this(0,0);
         this.id=id;
     }
