+++ src/main/java/org/trash_hunter/trashes/Can.java
@@ -21,7 +21,7 @@ public class Can extends Trash {
         appearanceRangeYSup= 700-this.height;
     }
     public Can (){this(0,0);}
-    public Can (long id){
+    public Can (int id){
         this(0,0);
         this.id=id;
     }
