+++ src/main/java/org/trash_hunter/trashes/OilContainer.java
@@ -21,7 +21,7 @@ public class OilContainer extends Trash {
         appearanceRangeYSup= 780-this.height;
     }
     public OilContainer(){this(0,0);}
-    public OilContainer (long id){
+    public OilContainer (int id){
         this(0,0);
         this.id=id;
     }
