+++ src/main/java/org/trash_hunter/trashes/PlasticBag.java
@@ -21,7 +21,7 @@ public class PlasticBag extends Trash {
         appearanceRangeYSup= 700-this.height;
     }
     public PlasticBag (){this(0,0);}
-    public PlasticBag (long id){
+    public PlasticBag (int id){
         this(0,0);
         this.id=id;
     }
