+++ src/main/java/org/trash_hunter/trashes/Tire.java
@@ -21,7 +21,7 @@ public class Tire extends Trash{
         appearanceRangeYSup= 780-this.height;
     }
     public Tire (){this(0,0);}
-    public Tire (long id){
+    public Tire (int id){
         this(0,0);
         this.id=id;
     }
